Move GET endpoint for a submission to new file, remove old files related to form submission
