+++ SocialCareCaseViewerApi.Tests/V1/UseCase/FormSubmissionUseCaseTests.cs
@@ -73,5 +73,28 @@ namespace SocialCareCaseViewerApi.Tests.V1.UseCase
             act.Should().Throw<PersonNotFoundException>()
                 .WithMessage($"Person with id {request.ResidentId} not found");
         }
+
+        [Test]
+        public void ExecuteGetByIdSuccessfully()
+        {
+            var submissionResponse = TestHelpers.CreateCaseSubmission();
+            _mockMongoGateway.Setup(x => x.LoadRecordById<CaseSubmission>(It.IsAny<string>(), submissionResponse.SubmissionId)).Returns(submissionResponse);
+
+            var response = _formSubmissionsUseCase.ExecuteGetById(submissionResponse.SubmissionId);
+
+            _mockMongoGateway.Verify(x => x.LoadRecordById<CaseSubmission>(It.IsAny<string>(), submissionResponse.SubmissionId), Times.Once);
+            response.Should().BeEquivalentTo(submissionResponse.ToDomain().ToResponse());
+        }
+
+        [Test]
+        public void ExecuteGetByIdShouldReturnNullIfNoCaseIsFound()
+        {
+            var nonExistentSubmissionId = new Guid();
+
+            var response = _formSubmissionsUseCase.ExecuteGetById(nonExistentSubmissionId);
+
+            _mockMongoGateway.Verify(x => x.LoadRecordById<CaseSubmission>(It.IsAny<string>(), nonExistentSubmissionId), Times.Once);
+            response.Should().BeNull();
+        }
     }
 }
