--- SocialCareCaseViewerApi.Tests/V1/UseCase/SubmissionUseCaseTests.cs
@@ -1,103 +1,0 @@
-using System;
-using FluentAssertions;
-using Moq;
-using NUnit.Framework;
-using SocialCareCaseViewerApi.Tests.V1.Helpers;
-using SocialCareCaseViewerApi.V1.Exceptions;
-using SocialCareCaseViewerApi.V1.Factories;
-using SocialCareCaseViewerApi.V1.Gateways;
-using SocialCareCaseViewerApi.V1.Infrastructure;
-using SocialCareCaseViewerApi.V1.UseCase;
-
-namespace SocialCareCaseViewerApi.Tests.V1.UseCase
-{
-    [TestFixture]
-    public class SubmissionUseCaseTests
-    {
-        private Mock<IDatabaseGateway> _mockDatabaseGateway;
-        private Mock<IMongoGateway> _mockMongoGateway;
-        private SubmissionsUseCase _submissionsUseCase;
-        private const string CollectionName = "resident-case-submissions";
-
-        [SetUp]
-        public void SetUp()
-        {
-            _mockDatabaseGateway = new Mock<IDatabaseGateway>();
-            _mockMongoGateway = new Mock<IMongoGateway>();
-            _submissionsUseCase = new SubmissionsUseCase(_mockDatabaseGateway.Object, _mockMongoGateway.Object);
-        }
-
-        [Test]
-        public void ExecutePostSuccessfully()
-        {
-            var request = TestHelpers.CreateCaseSubmissionRequest();
-            var worker = TestHelpers.CreateWorker();
-            var resident = TestHelpers.CreatePerson();
-
-            _mockDatabaseGateway.Setup(x => x.GetWorkerByEmail(request.CreatedBy)).Returns(worker);
-            _mockDatabaseGateway.Setup(x => x.GetPersonByMosaicId(request.ResidentId)).Returns(resident);
-            _mockMongoGateway.Setup(x => x.InsertRecord(It.IsAny<string>(), It.IsAny<CaseSubmission>()));
-
-            var (caseSubmissionResponse, caseSubmission) = _submissionsUseCase.ExecutePost(request);
-            var expectedResponse = TestHelpers.CreateCaseSubmission(SubmissionState.InProgress,
-                caseSubmission.CreatedAt, worker, resident, caseSubmission.SubmissionId, request.FormId);
-
-            caseSubmissionResponse.Should().BeEquivalentTo(expectedResponse.ToDomain().ToResponse());
-            _mockDatabaseGateway.Verify(x => x.GetWorkerByEmail(request.CreatedBy), Times.Once);
-            _mockDatabaseGateway.Verify(x => x.GetPersonByMosaicId(request.ResidentId), Times.Once);
-            _mockMongoGateway.Verify(x => x.InsertRecord(CollectionName, caseSubmission), Times.Once);
-        }
-
-        [Test]
-        public void ExecutePostThrowsAnErrorWhenNoWorkerFoundForRequest()
-        {
-            var request = TestHelpers.CreateCaseSubmissionRequest();
-            _mockDatabaseGateway.Setup(x => x.GetWorkerByEmail(request.CreatedBy));
-
-            Action act = () => _submissionsUseCase.ExecutePost(request);
-
-            act.Should().Throw<WorkerNotFoundException>()
-                .WithMessage($"Worker with email {request.CreatedBy} not found");
-        }
-
-        [Test]
-        public void ExecutePostThrowsAnErrorWhenNoResidentFoundForRequest()
-        {
-            var request = TestHelpers.CreateCaseSubmissionRequest();
-            var worker = TestHelpers.CreateWorker();
-            _mockDatabaseGateway.Setup(x => x.GetWorkerByEmail(request.CreatedBy)).Returns(worker);
-            _mockDatabaseGateway.Setup(x => x.GetPersonByMosaicId(request.ResidentId));
-
-            Action act = () => _submissionsUseCase.ExecutePost(request);
-
-            act.Should().Throw<PersonNotFoundException>()
-                .WithMessage($"Person with id {request.ResidentId} not found");
-        }
-
-        [Test]
-        public void ExecuteGetByIdSuccessfully()
-        {
-            var submissionResponse = TestHelpers.CreateCaseSubmission();
-            _mockMongoGateway.Setup(x => x.LoadRecordById<CaseSubmission>(It.IsAny<string>(), submissionResponse.SubmissionId)).Returns(submissionResponse);
-
-            var response = _submissionsUseCase.ExecuteGetById(submissionResponse.SubmissionId);
-
-            _mockMongoGateway.Verify(x => x.LoadRecordById<CaseSubmission>(It.IsAny<string>(), submissionResponse.SubmissionId), Times.Once);
-            response.Should().BeEquivalentTo(submissionResponse.ToDomain().ToResponse());
-        }
-
-        [Test]
-        public void ExecuteGetByIdShouldReturnNullIfNoCaseIsFound()
-        {
-            var nonExistentSubmissionId = new Guid();
-
-            var response = _submissionsUseCase.ExecuteGetById(nonExistentSubmissionId);
-
-            _mockMongoGateway.Verify(x => x.LoadRecordById<CaseSubmission>(It.IsAny<string>(), nonExistentSubmissionId), Times.Once);
-            response.Should().BeNull();
-        }
-
-
-
-    }
-}
