+++ SocialCareCaseViewerApi/V1/UseCase/Interfaces/IFormSubmissionsUseCase.cs
@@ -1,3 +1,4 @@
+using System;
 using SocialCareCaseViewerApi.V1.Boundary.Requests;
 using SocialCareCaseViewerApi.V1.Boundary.Response;
 using SocialCareCaseViewerApi.V1.Infrastructure;
@@ -7,5 +8,7 @@ namespace SocialCareCaseViewerApi.V1.UseCase.Interfaces
     public interface IFormSubmissionsUseCase
     {
         (CaseSubmissionResponse, CaseSubmission) ExecutePost(CreateCaseSubmissionRequest request);
+
+        CaseSubmissionResponse ExecuteGetById(Guid submissionId);
     }
 }
